refactor resolver
