--- compiler/src/core/assembler/index.ts
@@ -1,2 +1,0 @@
-export * from "./assemble.js";
-export type * from "./types.js";
