+++ compiler/src/core/builder/builders/double-builder.ts
@@ -1,4 +1,4 @@
-import type { Slice } from "#core/assembler/@";
+import type { Slice } from "#core/layout/@";
 import { Block } from "../block.js";
 import { Builder } from "./builder.js";
