+++ compiler/src/core/builder/size.ts
@@ -1,4 +1,4 @@
-import type { SongLayout } from "#core/assembler/@";
+import type { SongLayout } from "#core/layout/@";
 
 export type Size = {
 	width: number;
