+++ compiler/src/core/compile.ts
@@ -1,11 +1,11 @@
 import { UserError } from "#cli/error.js";
-import { assemble } from "#core/assembler/@";
 import { type Building, build, cachedBuilder } from "#core/builder/@";
-import { liveResolver, resolve } from "#core/resolver/@";
-import type { FileRef, JsonData } from "#schema/@";
+import { calculateLayout } from "#core/layout/@";
+import { type JsonString, liveResolver, resolve } from "#core/resolver/@";
+import type { FileRef } from "#schema/@";
 
-export function compile(src: FileRef | JsonData): Promise<Building> {
-	return resolve(src).then(assemble).then(build);
+export function compile(src: FileRef | JsonString): Promise<Building> {
+	return resolve(src).then(calculateLayout).then(build);
 }
 
 type Payload = Building | { error: string };
@@ -17,7 +17,7 @@ export async function* liveCompiler(
 	const build = cachedBuilder(options);
 	for await (const resolve of liveResolver(src, options)) {
 		const payload = await resolve()
-			.then(assemble)
+			.then(calculateLayout)
 			.then(build)
 			.catch((error) => {
 				if (!(error instanceof UserError)) {
