+++ compiler/src/core/layout/errors.ts
@@ -1,12 +1,48 @@
-import type { TickEvent } from "#core/resolver/@";
-import type { NoteEvent } from "./types.js";
+import { UserError } from "#cli/error.js";
+import type { IMeasure, NoteEvent, TickEvent } from "#core/resolver/@";
 
-type ValidateResult<T> = { value: T } | { error: string };
+export class ErrorTracker {
+	private readonly errors = new Map<string, Set<string>>();
+
+	registerError = (error: string | Error, measure: IMeasure) => {
+		const key = JSON.stringify(measure);
+		const message = typeof error === "string" ? error : error.message;
+
+		if (this.errors.has(key)) {
+			this.errors.get(key)!.add(message);
+		} else {
+			this.errors.set(key, new Set([message]));
+		}
+	};
+
+	validate() {
+		if (this.errors.size) {
+			return new UserError(this.formatErrorMessage());
+		}
+	}
+
+	private formatErrorMessage() {
+		const lines: string[] = [];
+
+		this.errors.forEach((errorSet, key) => {
+			const { bar, tick } = JSON.parse(key) as IMeasure;
+			lines.push(`ERROR @(${bar}, ${tick}):`);
+
+			errorSet.forEach((error) => {
+				lines.push(`  - ${error}`);
+			});
+
+			lines.push("");
+		});
+
+		return lines.slice(0, -1).join("\n");
+	}
+}
 
 export function validateConsistency<
 	E extends TickEvent.Voiced,
 	P extends string & keyof E,
->(events: E[], property: P): ValidateResult<E[P]> {
+>(events: E[], property: P): { value: E[P] } | { error: string } {
 	const propertyByVoice = new Map<string, { value: E[P]; voice: string }>();
 	events.forEach((event) => {
 		const { voice } = event;
