+++ compiler/src/core/resolver/components/index.ts
@@ -1,4 +1,3 @@
-export type { IMeasure } from "./measure.js";
-export type * from "./resolution.js";
-export * from "./song/song.js";
+export * from "./song.js";
 export type * from "./tick.js";
+export type { IMeasure } from "./utils/measure.js";
