+++ compiler/src/core/resolver/properties/duration.ts
