+++ compiler/src/core/resolver/properties/positional/division.ts
@@ -1,8 +1,8 @@
 import { times, zipWith } from "lodash";
 import { match, P } from "ts-pattern";
 import { assert, createIs } from "typia";
-import { resolveVariableValue } from "#core/resolver/duration.js";
 import type { Division as T_Division } from "#schema/@";
+import { resolveVariableValue } from "../duration.js";
 import { Positional } from "../positional.js";
 import type { ResolveType } from "../properties.js";
 import type { VariableTransformation } from "../variable.js";
