+++ compiler/src/core/resolver/properties/positional/trill.ts
@@ -1,5 +1,5 @@
-import { resolveDuration } from "#core/resolver/duration.js";
 import type { IPositional, Trill as T_Trill } from "#schema/@";
+import { resolveDuration } from "../duration.js";
 import { multiMap } from "../multi.js";
 import { Positional } from "../positional.js";
