--- compiler/src/core/validator/index.ts
@@ -1,2 +1,0 @@
-export * from "./song.js";
-export * from "./voice.js";
