--- compiler/src/types/schema/components/deferred.d.ts
@@ -1,8 +1,0 @@
-export type FileRef = `file://${string}`;
-
-export type JsonData = `json://${string}`;
-
-export type Deferred<
-	Data extends object,
-	Options extends { allowJson: boolean } = { allowJson: false },
-> = Data | FileRef | (Options["allowJson"] extends true ? JsonData : never);
