+++ compiler/src/types/schema/components/index.d.ts
@@ -1,4 +1,4 @@
-export type * from "./deferred.ts";
 export type * from "./note.ts";
+export type * from "./ref.js";
 export type * from "./song.ts";
 export type * from "./voice.ts";
