+++ compiler/src/types/schema/components/ref.d.ts
@@ -1,0 +1,1 @@
+export type FileRef = `file://${string}.yaml`;
