+++ compiler/src/types/schema/components/song.d.ts
@@ -1,14 +1,14 @@
 import type { tags } from "typia";
 import type { IGlobal, IProperties, TPosition } from "#schema/properties/@";
-import type { Deferred } from "./deferred.ts";
+import type { FileRef } from "./ref.js";
 import type { Notes, TValidate, Voice } from "./voice.ts";
 
-type VoiceGroup<T extends TValidate = TPosition> = Deferred<Voice<T>>[] &
+type VoiceGroup<T extends TValidate = TPosition> = (Voice<T> | FileRef)[] &
 	tags.MinItems<1>;
 
 export type VoiceEntry<T extends TValidate = TPosition> =
 	| null
-	| Deferred<Voice<T>>
+	| (Voice<T> | FileRef)
 	| VoiceGroup<T>;
 
 export type Voices<T extends TValidate = TPosition> = VoiceEntry<T>[] &
