+++ compiler/src/types/schema/components/voice.d.ts
@@ -1,7 +1,7 @@
 import type { BarLine } from "#schema/note/@";
 import type { IGlobal, IProperties, TPosition } from "#schema/properties/@";
-import type { Deferred } from "./deferred.ts";
 import type { Note } from "./note.ts";
+import type { FileRef } from "./ref.js";
 
 export type TValidate = TPosition | "lazy";
 
@@ -10,5 +10,5 @@ export type Notes<T extends TValidate = TPosition> = (T extends TPosition
 	: unknown)[];
 
 export type Voice<T extends TValidate = TPosition> = IGlobal<IProperties<T>> & {
-	notes: Deferred<Notes<T>>;
+	notes: Notes<T> | FileRef;
 };
